pyproject.toml
[project]
name = "firstagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.4",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
]
